+++ CHANGELOG.md
@@ -11,6 +11,9 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 ### Changed
+- `cache::Cache::cached_object` now returns `None` as soon as the intersection set of candidate
+  objects for the outputs is empty.  Before, that function computed the candidates of all outputs
+  before computing the intersection.
 
 ### Fixed